Initial Import.
This system was developed for PIRL/HiRISE DNS functions in
the Lunar and Planetary Lab at the University of Arizona. Thanks
to all involved in testing and implementing the system.

This import is taken from the current state of the HiRISE DNS scheme
with the HIRISE host/domain/network lists replaced by examples. In
time this example base should grow to help people quickly understand
how to use the system effectively.


git-svn-id: https://xmdns.svn.sourceforge.net/svnroot/xmdns@1 e49e35d3-a32e-4308-b9e8-b188a5c4e317